build: Remove automatic generation of manifest
I'm not sure it ever really made much sense to embed a Go program in an
artwork collection just to generate an XML file. I think that was a
higher barrier to contributing than just writing the XML by hand. The
script was also pretty inflexible regarding file names and formats.
Let's remove the complexity.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>